- Fix svg picture and add secondaryColor
