+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.USE_BIOMETRIC"/>
     <application
-        android:label="flutter"
+        android:label="flutter_note_mate"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
