+++ lib/core/theming/app_color.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 
 class AppColor {
   static const Color primaryColor = Color(0xFF252525);
-  
+
+  static const Color secondaryColor = Color(0xFFFFD523);
+
   static const Color grey = Color(0xFF606060);
 }
