+++ lib/core/widget/auth/text_form_field.dart
@@ -1,5 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
+import 'package:flutter_note_mate/core/theming/app_color.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_note_mate/core/theming/text_styles.dart';
 
@@ -17,7 +18,6 @@ class AppTextFormField extends StatelessWidget {
   final TextInputAction textInputAction;
   final Function(String?)? validator;
   final TextEditingController controller;
-  final double? width;
 
   const AppTextFormField({
     super.key,
@@ -29,7 +29,6 @@ class AppTextFormField extends StatelessWidget {
     required this.controller,
     this.readOnly = false,
     this.onTap,
-    this.width,
     this.keyboardType,
     required this.textInputAction,
     this.maxLine = 1,
@@ -39,71 +38,68 @@ class AppTextFormField extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      width: width,
-      child: Column(
-        children: [
-          TextFormField(
-            onTap: onTap,
-            onChanged: onChanged,
-            onFieldSubmitted: onFieldSubmitted,
-            maxLines: maxLine,
-            readOnly: readOnly,
-            validator: (value) {
-              if (validator != null) {
-                return validator!(value);
-              }
-              return null;
-            },
-            controller: controller,
-            keyboardType: keyboardType,
-            obscureText: isObscureText,
-            textInputAction: textInputAction,
-            style: AppTextStyles.font18MediumWhite,
-            decoration: InputDecoration(
-              fillColor: Colors.grey.withOpacity(0.1),
-              filled: true,
-              isDense: true,
-              hintStyle: AppTextStyles.font16Medium,
-              hintText: hintText,
-              prefixIcon: prefixIcon,
-              suffixIcon: suffixIcon,
-              contentPadding: EdgeInsets.symmetric(
-                horizontal: 20.w,
-                vertical: 20.h,
+    return Column(
+      children: [
+        TextFormField(
+          onTap: onTap,
+          onChanged: onChanged,
+          onFieldSubmitted: onFieldSubmitted,
+          maxLines: maxLine,
+          readOnly: readOnly,
+          validator: (value) {
+            if (validator != null) {
+              return validator!(value);
+            }
+            return null;
+          },
+          controller: controller,
+          keyboardType: keyboardType,
+          obscureText: isObscureText,
+          textInputAction: textInputAction,
+          style: AppTextStyles.font18MediumWhite,
+          decoration: InputDecoration(
+            fillColor: Colors.grey.withOpacity(0.1),
+            filled: true,
+            isDense: true,
+            hintStyle: AppTextStyles.font16Medium,
+            hintText: hintText,
+            prefixIcon: prefixIcon,
+            suffixIcon: suffixIcon,
+            contentPadding: EdgeInsets.symmetric(
+              horizontal: 20.w,
+              vertical: 20.h,
+            ),
+            focusedBorder: OutlineInputBorder(
+              borderSide: BorderSide(
+                color: AppColor.secondaryColor,
+                width: 1.3.w,
               ),
-              focusedBorder: OutlineInputBorder(
-                borderSide: BorderSide(
-                  color: Colors.green,
-                  width: 1.3.w,
-                ),
-                borderRadius: BorderRadius.circular(16.0),
+              borderRadius: BorderRadius.circular(16.0),
+            ),
+            errorBorder: OutlineInputBorder(
+              borderSide: const BorderSide(
+                color: Colors.red,
+                width: 1.3,
               ),
-              errorBorder: OutlineInputBorder(
-                borderSide: const BorderSide(
-                  color: Colors.red,
-                  width: 1.3,
-                ),
-                borderRadius: BorderRadius.circular(16.0),
+              borderRadius: BorderRadius.circular(16.0),
+            ),
+            focusedErrorBorder: OutlineInputBorder(
+              borderSide: const BorderSide(
+                color: Colors.red,
+                width: 1.3,
               ),
-              focusedErrorBorder: OutlineInputBorder(
-                borderSide: const BorderSide(
-                  color: Colors.red,
-                  width: 1.3,
-                ),
-                borderRadius: BorderRadius.circular(16.0),
+              borderRadius: BorderRadius.circular(16.0),
+            ),
+            enabledBorder: OutlineInputBorder(
+              borderSide: BorderSide(
+                color: const Color(0xFFEDEDED),
+                width: 1.3.w,
               ),
-              enabledBorder: OutlineInputBorder(
-                borderSide: BorderSide(
-                  color: const Color(0xFFEDEDED),
-                  width: 1.3.w,
-                ),
-                borderRadius: BorderRadius.circular(16.0),
-              ),
+              borderRadius: BorderRadius.circular(16.0),
             ),
           ),
-        ],
-      ),
+        ),
+      ],
     );
   }
 }
