+++ lib/features/note_veiw/view/edit_note/edit_note.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_note_mate/core/theming/app_color.dart';
 import 'package:flutter_note_mate/features/note_veiw/data/models/note_model.dart';
-import 'package:flutter_note_mate/features/note_veiw/view/widget/edit_note/edit_note_view_body.dart';
+import 'package:flutter_note_mate/features/note_veiw/view/edit_note/edit_note_view_body.dart';
 
 class EditeNoteView extends StatelessWidget {
   const EditeNoteView({super.key, required this.note});
