+++ lib/features/note_veiw/view/edit_note/edit_note_view_body.dart
@@ -6,7 +6,7 @@ import 'package:flutter_note_mate/core/theming/text_styles.dart';
 import 'package:flutter_note_mate/core/widget/auth/text_form_field.dart';
 import 'package:flutter_note_mate/features/note_veiw/logic/notes_cubit.dart';
 import 'package:flutter_note_mate/features/note_veiw/data/models/note_model.dart';
-import 'package:flutter_note_mate/features/note_veiw/view/widget/note_view/custom_app_bar.dart';
+import 'package:flutter_note_mate/features/note_veiw/view/note_view/custom_app_bar.dart';
 
 class EditNoteViewBody extends StatelessWidget {
   const EditNoteViewBody({super.key, required this.note});
