+++ lib/features/note_veiw/view/note_view/custom_icon.dart
