+++ lib/features/note_veiw/view/note_view/notes_view_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../logic/notes_cubit.dart';
+import '../../logic/notes_cubit.dart';
 import '../search_note/search_view.dart';
 import 'custom_app_bar.dart';
 import 'notes_list_view.dart';
@@ -29,7 +29,9 @@ class _NotesViewBodyState extends State<NotesViewBody> {
           CustomAppBar(
             title: 'Notes',
             perfixIcon: Icons.menu,
-            onPressedInPerfixIcon: () {},
+            onPressedInPerfixIcon: () {
+              Scaffold.of(context).openDrawer();
+            },
             suffixIcon: Icons.search,
             onPressedInSuffixIcon: () {
               Navigator.push(context, MaterialPageRoute(builder: (context) {
