+++ lib/features/note_veiw/view/notes_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_note_mate/core/theming/app_color.dart';
-import 'package:flutter_note_mate/features/note_veiw/view/widget/add_note/add_note_view.dart';
-import 'package:flutter_note_mate/features/note_veiw/view/widget/note_view/notes_view_body.dart';
+import 'package:flutter_note_mate/features/note_veiw/view/add_note/add_note_view.dart';
+import 'package:flutter_note_mate/features/note_veiw/view/note_view/notes_view_body.dart';
 
 class NotesView extends StatelessWidget {
   const NotesView({super.key});
@@ -10,6 +10,7 @@ class NotesView extends StatelessWidget {
   Widget build(BuildContext context) {
     return const Scaffold(
       backgroundColor: AppColor.primaryColor,
+      drawer: Drawer(),
       resizeToAvoidBottomInset: false,
       floatingActionButton: AddNoteView(),
       body: NotesViewBody(),
