+++ lib/features/pin_code/ui/pin_code.dart
@@ -30,7 +30,7 @@ class PinCode extends StatelessWidget {
                 SizedBox(
                   height: 250,
                   child: SvgPicture.asset(
-                    AppAssests.fingerprint,
+                    AppAssests.pinCode,
                   ),
                 ),
                 Padding(
