+++ lib/note_mate.dart
@@ -52,7 +52,7 @@ Future<String> _checkIfUserIsLoggedIn() async {
   User? user = auth.currentUser;
   if (user != null) {
     // User is signed in
-    return Routes.pinCode;
+    return Routes.homeScreen;
   } else {
     // No user is signed in
     return Routes.loginScreen;
